Completed and tested the computeProfitLoss function in PurchaseSaleRecord

+++ PurchaseSaleRecord.h
@@ -22,13 +22,13 @@ public:
 
     std::string getCompany() const { return _company; }
 
-    int getPurchaseDate() const { return _purchaseDate; }
+    double getPurchaseDate() const { return _purchaseDate; }
 
     double getPurchasePrice() const { return _purchasePrice; }
 
     double getPurchaseFee() const { return _purchaseFee; }
 
-    int getSaleDate() const { return _saleDate; }
+    double getSaleDate() const { return _saleDate; }
 
     double getSalePrice() const { return _salePrice; }
 
@@ -36,11 +36,15 @@ public:
 
 private:
     std::string _company;
-    int _numShares;
-    int _purchaseDate;
-    int _purchasePrice;
-    int _purchaseFee;
-    int _saleDate;
+
+    /* The following data members are stored as doubles rather than ints due to the fact that
+     * the integers involved have the potential to be very large. Better safe than sorry.*/
+    // TODO: update UML diagram with new double values
+    double _numShares;
+    double _purchaseDate;
+    double _purchasePrice;
+    double _purchaseFee;
+    double _saleDate;
     double _salePrice;
     double _saleFee;
 };
